requirements.txt
autopep8==1.3.3
boto3==1.4.7
coverage==4.4.1
enum34==1.1.6
flake8>=2.5.1,<2.6
simplejson>=3.12.2
locustio==0.8.1
mock==2.0.0
nose==1.3.7
requests>=2.18.1
requests_mock>=1.3.0
wget==3.2
